ref(video-quality): Move all related code.
