+++ react/features/video-quality/index.js
@@ -1,4 +1,5 @@
 export * from './components';
 export * from './actions';
+export * from './actionTypes';
 
 import './reducer';
